Version 1.03 of the Mackie C4 Script - Moved elements pushEncoder and button to a separate file. - Moved Mackie C4 Surface to a separate file. - Moved eqPage to a separate file. - Swapped gain and frequency rows to align with cubase GUI. - Removed LED clearing for bank prev and channel prev as it is not needed.  Only needed for next. - Updated documentation for V 1.03
